+++ packages/playground/package.json
@@ -4,7 +4,7 @@
   "private": "true",
   "scripts": {
     "dev": "next",
-    "export": "next build && next export"
+    "build": "next build"
   },
   "dependencies": {
     "@code-hike/mdx": "^0.3.0-next.0",
